add backend test instead of separate test for sqlite and postgresql
