+++ internal/backends/collection_test.go
@@ -132,3 +132,44 @@ func TestCollectionUpdateAll(t *testing.T) {
 		})
 	}
 }
+
+func TestCollectionStats(t *testing.T) {
+	t.Parallel()
+
+	ctx := conninfo.Ctx(testutil.Ctx(t), conninfo.New())
+
+	for _, b := range testBackends(t) {
+		b := b
+		t.Run(b.Name(), func(t *testing.T) {
+			t.Parallel()
+
+			db, err := b.Database(testutil.DatabaseName(t))
+			require.NoError(t, err)
+
+			cNames := []string{"collectionOne", "collectionTwo"}
+			for _, cName := range cNames {
+				err = db.CreateCollection(ctx, &backends.CreateCollectionParams{Name: cName})
+				require.NoError(t, err)
+			}
+
+			c, err := db.Collection(cNames[0])
+			require.NoError(t, err)
+
+			_, err = c.InsertAll(ctx, &backends.InsertAllParams{
+				Docs: []*types.Document{must.NotFail(types.NewDocument("_id", types.NewObjectID()))},
+			})
+			require.NoError(t, err)
+
+			dbStatsRes, err := db.Stats(ctx, new(backends.DatabaseStatsParams))
+			require.NoError(t, err)
+
+			res, err := c.Stats(ctx, new(backends.CollectionStatsParams))
+			require.NoError(t, err)
+			require.NotZero(t, res.SizeTotal)
+			require.Less(t, res.SizeTotal, dbStatsRes.SizeTotal)
+			require.NotZero(t, res.SizeCollection)
+			require.Less(t, res.SizeCollection, dbStatsRes.SizeCollections)
+			require.Equal(t, res.CountObjects, int64(1))
+		})
+	}
+}
