+++ internal/backends/sqlite/collection_test.go
@@ -21,7 +21,6 @@ import (
 
 	"github.com/FerretDB/FerretDB/internal/backends"
 	"github.com/FerretDB/FerretDB/internal/types"
-	"github.com/FerretDB/FerretDB/internal/util/must"
 	"github.com/FerretDB/FerretDB/internal/util/state"
 	"github.com/FerretDB/FerretDB/internal/util/testutil"
 )
@@ -55,44 +54,4 @@ func TestInsert(t *testing.T) {
 		Docs: []*types.Document{doc},
 	})
 	require.True(t, backends.ErrorCodeIs(err, backends.ErrorCodeInsertDuplicateID))
-}
-
-func TestCollectionStats(t *testing.T) {
-	t.Parallel()
-	ctx := testutil.Ctx(t)
-
-	sp, err := state.NewProvider("")
-	require.NoError(t, err)
-
-	b, err := NewBackend(&NewBackendParams{URI: "file:./?mode=memory", L: testutil.Logger(t), P: sp})
-	require.NoError(t, err)
-	t.Cleanup(b.Close)
-
-	db, err := b.Database(testutil.DatabaseName(t))
-	require.NoError(t, err)
-
-	cNames := []string{"collectionOne", "collectionTwo"}
-	for _, cName := range cNames {
-		err = db.CreateCollection(ctx, &backends.CreateCollectionParams{Name: cName})
-		require.NoError(t, err)
-	}
-
-	c, err := db.Collection(cNames[0])
-	require.NoError(t, err)
-
-	_, err = c.InsertAll(ctx, &backends.InsertAllParams{
-		Docs: []*types.Document{must.NotFail(types.NewDocument("_id", types.NewObjectID()))},
-	})
-	require.NoError(t, err)
-
-	dbStatsRes, err := db.Stats(ctx, new(backends.DatabaseStatsParams))
-	require.NoError(t, err)
-
-	res, err := c.Stats(ctx, new(backends.CollectionStatsParams))
-	require.NoError(t, err)
-	require.NotZero(t, res.SizeTotal)
-	require.Less(t, res.SizeTotal, dbStatsRes.SizeTotal)
-	require.NotZero(t, res.SizeCollection)
-	require.Less(t, res.SizeCollection, dbStatsRes.SizeCollections)
-	require.Equal(t, res.CountObjects, int64(1))
 }
